+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Names of parameters and types in error messages are now reported inside backticks to make them more distinguishable from the rest of the message 
 - Raise an error when trying to show an animation whose frames are not all of type `Graphic`
+- Rendered text now includes leading and trailing glyphs without outline (such as spaces)
 
 ### Fixed